refactor(naming): rename function to match functionallity
