+++ lib/index.js
@@ -15,7 +15,7 @@ import {
   writeErrorLogFile
 } from 'contentful-batch-libs/dist/utils/logging'
 
-import getTransformedDestinationResponse from './get-transformed-destination-response'
+import getDestinationResponse from './get-destination-response'
 import parseOptions from './parseOptions'
 
 function createListrOptions (options) {
@@ -72,7 +72,7 @@ export default function runContentfulImport (params) {
       task: (ctx) => {
         return Promise.props({
           source: options.content,
-          destination: getTransformedDestinationResponse({
+          destination: getDestinationResponse({
             managementClient: ctx.clients.destination.management,
             spaceId: ctx.clients.destination.spaceId,
             sourceResponse: options.content,
@@ -89,13 +89,9 @@ export default function runContentfulImport (params) {
     {
       title: 'Apply transformations to source data',
       task: (ctx) => {
-        return Promise.props({
-          source: transformSpace(ctx.source, ctx.destination),
-          destination: ctx.destination
-        })
-          .then(({source, destination}) => {
+        return transformSpace(ctx.source, ctx.destination)
+          .then((source) => {
             ctx.source = source
-            ctx.destination = destination
           })
       }
     },
